+++ homework05/src/main/java/ru/otus/homework05/service/BookServiceImpl.java
@@ -22,7 +22,7 @@ public class BookServiceImpl implements BookService {
     }
 
     @Override
-    public void insert(int bookId, String bookName, int authorId, int genreId) {
+    public void insert(long bookId, String bookName, long authorId, long genreId) {
         Author author = authorService.getById(authorId);
         Genre genre = genreService.getById(genreId);
         Book book = new Book(bookId, bookName, author, genre);
@@ -30,12 +30,12 @@ public class BookServiceImpl implements BookService {
     }
 
     @Override
-    public void delete(int bookId) {
+    public void delete(long bookId) {
         bookDaoJdbc.deleteById(bookId);
     }
 
     @Override
-    public void update(int bookId, String bookName, int authorId, int genreId) {
+    public void update(long bookId, String bookName, long authorId, long genreId) {
         Author author = authorService.getById(authorId);
         Genre genre = genreService.getById(genreId);
         Book book = new Book(bookId, bookName, author, genre);
